Add assertations

+++ src/main/java/fitlogger/profile/UserProfile.java
@@ -2,8 +2,8 @@ package fitlogger.profile;
 
 public class UserProfile {
     private String name;
-    private float weight;
-    private float height;
+    private double weight;
+    private double height;
 
     public UserProfile() {
         name = null;
@@ -19,19 +19,19 @@ public class UserProfile {
         this.name = name;
     }
 
-    public float getWeight() {
+    public double getWeight() {
         return weight;
     }
 
-    public void setWeight(float weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
 
-    public float getHeight() {
+    public double getHeight() {
         return height;
     }
 
-    public void setHeight(float height) {
+    public void setHeight(double height) {
         this.height = height;
     }
 }
